+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.0.8
 * Bug Fixes: Fixed issue with startimagestream on macos 14.0
 * Added ability to turn light on or off
+* Added ability to change camera orientation
+* Added ability to zoom on imageStream
 * Removed deviceId requirment from setFocusPoint
 ## 0.0.7+2
 * Updated README.md